Consolidate web progress appData modules
