+++ apps/web/src/appData/progress/reviewProgressBadge.ts
@@ -1,8 +1,8 @@
 import { useMemo } from "react";
-import type { ProgressSummarySnapshot, ProgressSummarySourceState, ReviewProgressBadgeState } from "../types";
+import type { ProgressSummarySnapshot, ProgressSummarySourceState, ReviewProgressBadgeState } from "../../types";
 import { useProgressInvalidationState } from "./progressInvalidation";
 import { useProgressSource } from "./progressSource";
-import { useAppData } from "./provider";
+import { useAppData } from "../provider";
 
 const EMPTY_REVIEW_PROGRESS_BADGE_STATE: ReviewProgressBadgeState = {
   streakDays: 0,
