--- apps/web/src/appData/progressSource.ts
@@ -1,7 +1,0 @@
-export {
-  buildProgressSeriesInput,
-  parseLocalDate,
-  shiftLocalDate,
-} from "../progress/progressDates";
-export { buildProgressScopeKey } from "./progress/progressScope";
-export { useProgressSource } from "./progress/useProgressSource";
